+++ broker/src/main/java/com/flipkart/vbroker/services/TopicService.java
@@ -1,7 +1,7 @@
 package com.flipkart.vbroker.services;
 
 import com.flipkart.vbroker.core.TopicPartition;
-import com.flipkart.vbroker.entities.Topic;
+import com.flipkart.vbroker.wrappers.Topic;
 
 import java.util.List;
 import java.util.concurrent.CompletionStage;
@@ -12,7 +12,7 @@ public interface TopicService {
 
     // public void createTopicPartition(Topic topic, TopicPartition topicPartition);
 
-    public CompletionStage<TopicPartition> getTopicPartition(Topic topic, short topicPartitionId);
+    public CompletionStage<TopicPartition> getTopicPartition(Topic topic, int topicPartitionId);
 
     /**
      * check if the topic is present
@@ -22,7 +22,7 @@ public interface TopicService {
      */
     public CompletionStage<Boolean> isTopicPresent(short topicId);
 
-    public CompletionStage<Topic> getTopic(short topicId);
+    public CompletionStage<Topic> getTopic(int topicId);
 
     public List<TopicPartition> getPartitions(Topic topic);
