+++ client/src/main/java/com/flipkart/vbroker/client/ProducerRecord.java
@@ -15,11 +15,11 @@ public class ProducerRecord {
     private byte crc;
     private byte version;
     private int seqNo;
-    private short topicId;
+    private int topicId;
     private int attributes;
     private String httpUri;
     private HttpMethod httpMethod;
-    private short callbackTopicId;
+    private int callbackTopicId;
     private String callbackHttpUri;
     private HttpMethod callbackHttpMethod;
     private Map<String, String> headers;
