+++ client/src/main/java/com/flipkart/vbroker/handlers/ProduceResponseHandler.java
@@ -1,9 +1,9 @@
 package com.flipkart.vbroker.handlers;
 
-import com.flipkart.vbroker.entities.ProduceResponse;
-import com.flipkart.vbroker.entities.TopicPartitionProduceResponse;
-import com.flipkart.vbroker.entities.TopicProduceResponse;
-import com.flipkart.vbroker.entities.VResponse;
+import com.flipkart.vbroker.flatbuf.ProduceResponse;
+import com.flipkart.vbroker.flatbuf.TopicPartitionProduceResponse;
+import com.flipkart.vbroker.flatbuf.TopicProduceResponse;
+import com.flipkart.vbroker.flatbuf.VResponse;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -23,7 +23,7 @@ public class ProduceResponseHandler implements ResponseHandler {
 
         for (int i = 0; i < produceResponse.topicResponsesLength(); i++) {
             TopicProduceResponse topicProduceResponse = produceResponse.topicResponses(i);
-            short topicId = topicProduceResponse.topicId();
+            int topicId = topicProduceResponse.topicId();
             log.info("Handling ProduceResponse for topic {} with {} partition responses", topicId, topicProduceResponse.partitionResponsesLength());
             for (int j = 0; j < topicProduceResponse.partitionResponsesLength(); j++) {
                 TopicPartitionProduceResponse partitionProduceResponse = topicProduceResponse.partitionResponses(j);
